Login Working
Login is working a session and cookie is created, navabar is updated
along with the username of who just logged in. Need to do check login on
each page to update navbar accordingly and get the username.  Maybe Have
to send the cookie to Authenticate.ensureLoggedIn and send username back
if logged in.

Logout button is shown when logged in on main page and will try to do a
logout post request to the server. Still have to add back to reshow
login after logout.