Basic navigation between subsaiddits and posts
Navigation to a subsaiddit working, a url is generated of the form
/api/subsaidditname/ which is the front page of the subsaiddit. And for
the post is /api/subsaiditname/postTitle. Next I will make the HTML
update according to the given url subsaiddit to update the title and
posts it shows and on posts to get the title of subsaiddit, post title
and actual text of the post.