Login and Logout
Login and logout working and is persistent between pages.

Still to do:
Create a logged-in "front-page" for the user. (A user page) That shows
its top-voted posts from the subscribed subsaiddits.

Create form to delete post.  (I think this sould be inside the post
itself and only show when the user create the post)

Create forms to populate tables. Not sure what this is. Maybe create new
subsaiddits, and create comments or so.