Posts
Posts are now displayed correctly when selected.

Still to add for requirements:

Delete form. I was thinking of creating a delete post link on the
nabber in my page that will display the posts that belongs to that user
and have the option of clicking a delete button.

Add post. Will work based on the url of the subsaiddit plus the logged
user to create the post.

Also to display the rating of the post, it will show the rating of the
first one. Have to fix the SQL so that id adds up all the rating
corresponding to a post and display 0 if there is no entry in the
post_rating.